# apps/api/generated/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qafiyah API",
...
    "contact": {
      "name": "Qafiyah",
      "url": "https://qafiyah.com"
    },
    "license": {
      "name": "CC0 1.0 (data)",
      "identifier": "CC0-1.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "paths": {
...
    "/poems": {
      "get": {
        "tags": [
          "poems"
        ],
        "description": "A page of 30 poems in catalog order, oldest entries first, holding primary readings only (a poem's alternate readings are listed on the poem). Filter by poet, era, theme, meter, rhyme, and collection: values of one filter combine with OR and different filters with AND, e.g. `?poet=PAKT&meter=altawil&meter=alkamil`. A slug that matches nothing gives an empty page rather than an error, and unknown query params are ignored.",
        "operationId": "poems.list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number as a 1-based integer string, 30 poems a page. Minimum 1.",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]*$"
            },
            "example": "1"
          },
          {
            "name": "poet",
            "in": "query",
            "description": "Filter by poet. Repeatable, e.g. `?poet=PAKT`. Values are `slug` from GET /poets.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/FourLetterSlug"
              },
              "default": [],
              "maxItems": 100
            },
            "example": [
              "PAKT"
            ]
          },
          {
            "name": "era",
            "in": "query",
            "description": "Filter by era. Repeatable, e.g. `?era=jahili`. Values are `slug` from GET /eras.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TransliteratedSlug"
              },
              "default": [],
              "maxItems": 100
            },
            "example": [
              "jahili"
            ]
          },
          {
            "name": "theme",
            "in": "query",
            "description": "Filter by theme. Repeatable, e.g. `?theme=alhikma`. Values are `slug` from GET /themes.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TransliteratedSlug"
              },
              "default": [],
              "maxItems": 100
            },
            "example": [
              "alhikma"
            ]
          },
          {
            "name": "meter",
            "in": "query",
            "description": "Filter by meter. Repeatable, e.g. `?meter=altawil`. Values are `slug` from GET /meters.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TransliteratedSlug"
              },
              "default": [],
              "maxItems": 100
            },
            "example": [
              "altawil"
            ]
          },
          {
            "name": "rhyme",
            "in": "query",
            "description": "Filter by rhyme. Repeatable, e.g. `?rhyme=meem`. Values are `slug` from GET /rhymes.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TransliteratedSlug"
              },
              "default": [],
              "maxItems": 100
            },
            "example": [
              "meem"
            ]
          },
          {
            "name": "collection",
            "in": "query",
            "description": "Filter by collection. Repeatable, e.g. `?collection=almuallaqat`. Values are `slug` from GET /collections.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TransliteratedSlug"
              },
              "default": [],
              "maxItems": 100
            },
            "example": [
              "almuallaqat"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "A page of poems with pagination metadata.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Send back in If-None-Match to get 304 while the data is unchanged"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_PoemListItem"
                }
              }
            }
          },
          "304": {
            "description": "Not modified: the If-None-Match ETag still matches, so the body is empty.",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable: the database or search index did not answer in time",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds to wait before retrying"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/poems/slugs": {
      "get": {
        "tags": [
          "poems"
        ],
        "description": "The slug of every primary poem, 45,000 a page in slug order, for sitemaps and incremental crawling.",
        "operationId": "poems.listSlugs",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number as a 1-based integer string, 45,000 slugs a page. Minimum 1.",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]*$"
            },
            "example": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of poem slugs.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Send back in If-None-Match to get 304 while the data is unchanged"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_FourLetterSlug"
                }
              }
            }
          },
          "304": {
            "description": "Not modified: the If-None-Match ETag still matches, so the body is empty.",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable: the database or search index did not answer in time",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds to wait before retrying"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/poets/slugs": {
      "get": {
        "tags": [
          "poets"
        ],
        "description": "The slug of every poet with at least one poem, with an avatar flag, 45,000 a page in slug order, for sitemaps and incremental crawling.",
        "operationId": "poets.listSlugs",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number as a 1-based integer string, 45,000 slugs a page. Minimum 1.",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]*$"
            },
            "example": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of poet slugs, each flagged with whether the poet has an avatar.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Send back in If-None-Match to get 304 while the data is unchanged"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_PoetSlugEntry"
                }
              }
            }
          },
          "304": {
            "description": "Not modified: the If-None-Match ETag still matches, so the body is empty.",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable: the database or search index did not answer in time",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds to wait before retrying"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EraRef": {
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The era's name in Arabic.",
            "example": "جاهلي"
          },
          "slug": {
            "type": "string",
            "description": "The era's slug, for `GET /eras/{slug}`.",
            "example": "jahili",
            "pattern": "^[a-z][a-z-]*$"
          }
        }
      },
...
      "FourLetterSlug": {
        "type": "string",
        "pattern": "^[a-zA-Z]{4}$"
      },
...
      "ListEnvelope_PoemListItem": {
        "type": "object",
        "description": "A page of items under `data`, with its place in the whole list under `pagination`.",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "title",
                "slug",
                "poet",
                "meter"
              ],
              "properties": {
                "era": {
                  "$ref": "#/components/schemas/EraRef",
                  "description": "The poem's era, which is its poet's. Included in `relatedPoems`, left out in `GET /poems` lists."
                },
                "meter": {
                  "$ref": "#/components/schemas/MeterRef",
                  "description": "The poem's meter."
                },
                "poet": {
                  "$ref": "#/components/schemas/PoetRef",
                  "description": "The poet."
                },
                "slug": {
                  "type": "string",
                  "description": "The poem's four-letter, case-sensitive slug, for `GET /poems/{slug}`.",
                  "example": "gnNg",
                  "pattern": "^[a-zA-Z]{4}$"
                },
                "title": {
                  "type": "string",
                  "description": "The poem's title, without diacritics. Most poems are titled by their opening half-line.",
                  "example": "أمن أم أوفى دمنة لم تكلم"
                }
              }
            },
            "description": "The items of this page."
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination",
            "description": "Where this page sits in the list."
          }
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "instance",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable machine-readable code, such as `NOT_FOUND`, `BAD_REQUEST`, `TOO_MANY_REQUESTS`, or `SERVICE_UNAVAILABLE`.",
            "example": "NOT_FOUND"
          },
          "detail": {
            "type": "string",
            "description": "What went wrong with this request, in English.",
            "example": "Meter not found"
          },
          "instance": {
            "type": "string",
            "description": "The path of the request that failed.",
            "example": "/v1/meters/zzzz"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code.",
            "example": 404
          },
          "title": {
            "type": "string",
            "description": "A short summary of the kind of error.",
            "example": "Resource not found"
          },
          "type": {
            "type": "string",
            "format": "uri",
            "description": "A URI identifying the kind of error. It is an identifier and does not resolve to a page.",
            "example": "https://qafiyah.com/errors/not-found"
          }
        }
      },
...
      "TransliteratedSlug": {
        "type": "string",
        "pattern": "^[a-z][a-z-]*$"
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Optional. A free key from https://qafiyah.com/developers raises the rate limit. A key the API does not recognize counts as no key."
      }
    }
  },
  "security": [
    {},
    {
      "apiKey": []
    }
  ],
...
}